+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -2,6 +2,7 @@
 
 module Test.Cardano.Ledger.Api.Arbitrary () where
 
+import Test.Cardano.Ledger.Dijkstra.Arbitrary ()
 import Cardano.Ledger.Api.State.Query (MemberStatus)
 import Test.Cardano.Ledger.Common